// src/types/locationTypes.ts
// Generated by https://quicktype.io

export interface LocationInfo {
  coords: Coords;
  extras: Extras;
  mocked: boolean;
  timestamp: number;
}

export interface Coords {
  accuracy: number;
  altitude: number;
  heading: number;
  latitude: number;
  longitude: number;
  speed: number;
}

export interface Extras {
  networkLocationType: string;
}